Simplify to single .server.ts convention, remove shared/ directory
Move utilities from shared/subscriptionItemUtils.ts back to subscriptionItemMethods.ts.
Now that the circular dependency is broken by .server.ts, the shared/ directory
is no longer needed.
